+++ demo/example.js
@@ -1,13 +1,13 @@
-const prettify = require('../index.js');
+const pretty = require('../index.js');
 
 const value = {
 	aNumber: 1,
 	aString: '2',
-	aRegExp: /abcd/gi,
+	aRegExp: /ab\r\ncd/gi,
 	anObject: {
 		aNull: null,
 		aBoolean: false,
-		aBoolean2: true,
+		aBooleanToo: true,
 		twoPromises: [Promise.resolve(1), new Promise(() => {})],
 		anUndefined: undefined,
 		$anArray: [1, new Date()],
@@ -15,14 +15,15 @@ const value = {
 		'0Infinity': 5 / 0,
 		'a NaN': NaN,
 		'negative Infinity': -5 / 0,
-		_aMultilineString: 'diner\nstays\r\nopen" `ok`?',
 		aString: 'say "hello"!',
+		aUnicodeString: 'Ich ♥ Bücher; foo 𝌆 bar 🏁 baz',
+		_aMultilineString: 'diner\nstays\r\n"open".\t`ok`?',
 	},
 	aSymbol: Symbol('six'),
 	aBigInt: BigInt(Number.MAX_SAFE_INTEGER) * 2n,
 	anError: new Error('foobar'),
-	aURL: new URL('https://example.com/abc?d=4'),
-	aURLSearchParams: new URLSearchParams({ d: 4, e: 5 }),
+	aURL: new URL('https://example.com/abc/🛹/?d=4'),
+	aURLSearchParams: new URLSearchParams({ d: 4, e: '🤿' }),
 	aSet: new Set(['a', 'b']),
 	aMap: new Map([
 		['c', 3],
@@ -32,11 +33,13 @@ const value = {
 	aWeakMap: new WeakMap(),
 	anEmptyArray: [],
 	anEmptyObject: {},
-	aBuffer: Buffer.from([42, 422, 4222, 4242, 42424]),
+	aBuffer: Buffer.from([42, 4, 1, 20]),
+	aBufferToo: Buffer.from('🪂⛳️'),
 };
 value.aCircularArray = value.anObject.$anArray;
 value.aCircularObject = value.anEmptyObject;
 value.aCircularSet = value.aSet;
 
 console.log('example:');
-prettify.log(value);
+// pretty.options.preferBackticks = false;
+pretty.log(value);