+++ src/options/options.js
@@ -4,6 +4,9 @@ const options = new Store({
 	showFunctionBody: false,
 	maxStringLength: 1024 * 4,
 	maxArrayItems: 100,
+	preferBackticks: true,
+	// allowed: single, double, backtick
+	quoteStyle: 'double',
 });
 
 module.exports = options;