+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## v1.1.0 - 2 Apr 2020
 
 - Full unit tests
-- Configurable indentation, labels, colors 
+- Properly escape complex strings using jsesc
+- Configurable indentation, labels, colors
+- Color functions from ansi-colors
 
 ## v1.0.2 - 25 Feb 2020